Allow for time adjustments with +/-
I much prefer this solution to the old one. Parsing times was a mistake.

I wonder if rustc realizes that I only ever pass 5 or -5 as arguments to
that function and then creates two `time::Duration`s for 5 and -5
minutes respectively to avoid the allocations on each keypress.

Also, I had to satisfy the OCD meme by rounding all adjusted times to %5.
It would be too annoying to change a number without being able to always
set it to a “nice” number.